+++ packages/dso-toolkit/src/components/info-button/info-button.models.ts
@@ -3,7 +3,6 @@ export type InfoButtonTooltipPlacement = "top" | "bottom" | "left" | "right";
 export interface InfoButton<TemplateFnReturnType> {
   active?: boolean;
   toggletipPlacement?: InfoButtonTooltipPlacement;
-  secondary?: boolean;
   label?: string;
   dsoToggle?: (e: CustomEvent<InfoButtonToggleEvent>) => void;
   children?: TemplateFnReturnType;